+++ docs/src/app/pages/components/Chip/ExampleChipDefault.js
@@ -1,0 +1,8 @@
+import React from 'react'
+import Chip from 'react-ions/lib/components/Chip'
+
+const ExampleChipDefault = () => (
+  <Chip text='Default' />
+)
+
+export default ExampleChipDefault