+++ docs/src/app/pages/components/Chip/ExampleChipSmaller.js
@@ -1,0 +1,8 @@
+import React from 'react'
+import Chip from 'react-ions/lib/components/Chip'
+
+const ExampleChipSmaller = () => (
+  <Chip text='High Contrast' size='smaller' color='neutral-1' />
+)
+
+export default ExampleChipSmaller