+++ docs/src/app/pages/components/Input/ExampleInputSuffix.js
@@ -1,9 +1,8 @@
 import React from 'react'
 import Input from 'react-ions/lib/components/Input'
-import style from './style.scss'
 
 const ExampleInputSuffix = () => (
-  <Input suffix='days' optClass={style['prefix-suffix']} />
+  <Input suffix='days' width='100px' />
 )
 
 export default ExampleInputSuffix