+++ docs/src/app/routes.js
@@ -16,6 +16,7 @@ import ColorsPage from './pages/foundations/Colors/Page'
 import ColorPickerPage from './pages/components/ColorPicker/Page'
 import DatePickerPage from './pages/components/DatePicker/Page'
 import CheckboxPage from './pages/components/Checkbox/Page'
+import ChipPage from './pages/components/Chip/Page'
 import DropdownPage from './pages/components/Dropdown/Page'
 import HomePage from './pages/Home/Page'
 import IconographyPage from './pages/foundations/Iconography/Page'
@@ -31,6 +32,7 @@ import InputListPage from './pages/components/InputList/Page'
 import InlineEditPage from './pages/components/InlineEdit/Page'
 import MultiSelectPage from './pages/components/MultiSelect/Page'
 import TooltipPage from './pages/components/Tooltip/Page'
+import PopoverPage from './pages/components/Popover/Page'
 import ProgressBarPage from './pages/components/ProgressBar/Page'
 import PanelGroupPage from './pages/components/PanelGroup/Page'
 import SelectFieldPage from './pages/components/SelectField/Page'
@@ -63,6 +65,7 @@ const Routes = (
       <Route path='button-group' title='Button Group' component={ButtonGroupPage} />
       <Route path='buttons' title='Buttons' component={ButtonsPage} />
       <Route path='checkbox' title='Checkbox' component={CheckboxPage} />
+      <Route path='chip' title='Chip' component={ChipPage} />
       <Route path='color-picker' title='Color Picker' component={ColorPickerPage} />
       <Route path='date-picker' title='Date Picker' component={DatePickerPage} />
       <Route path='file-upload' title='File Upload' component={FileUploadPage} />
@@ -74,8 +77,9 @@ const Routes = (
       <Route path='modal' title='Modal' component={ModalPage} />
       <Route path='multi-select' title='Multi Select' component={MultiSelectPage} />
       <Route path='nav' title='Nav' component={NavPage} />
-      <Route path='progress-bar' title='Progress Bar' component={ProgressBarPage} />
       <Route path='panel-group' title='Panel Group' component={PanelGroupPage} />
+      <Route path='popover' title='Popover' component={PopoverPage} />
+      <Route path='progress-bar' title='Progress Bar' component={ProgressBarPage} />
       <Route path='select-field' title='Select Field' component={SelectFieldPage} />
       <Route path='sortable-list' title='Sortable List' component={SortableListPage} />
       <Route path='spinner' title='Spinner' component={SpinnerPage} />
@@ -92,4 +96,4 @@ const Routes = (
   </Route>
 )
 
-export default Routes;
+export default Routes